refactor: changed files component "Map" for MobX
